# docs/swagger.yaml
basePath: /
definitions:
...
  meals.MealEntry:
    properties:
      created_at:
        type: string
      effective_name:
        type: string
      effective_nutriments_per_100g:
        $ref: '#/definitions/meals.Nutriments'
      id:
        type: string
      logged_at:
        type: string
      meal_type:
        $ref: '#/definitions/meals.MealType'
      note:
        type: string
      product_id:
        type: string
      quantity_g:
        type: number
      updated_at:
        type: string
      workout_id:
        type: string
    type: object
  meals.MealType:
    enum:
    - breakfast
    - lunch
    - dinner
    - snack
    type: string
    x-enum-varnames:
    - Breakfast
    - Lunch
    - Dinner
    - Snack
  meals.Nutriments:
    properties:
      calcium_mg:
        type: number
      carbs_g:
        type: number
      fat_g:
        type: number
      fiber_g:
        type: number
      iron_mg:
        type: number
      kcal:
        type: number
      magnesium_mg:
        type: number
      potassium_mg:
        type: number
      protein_g:
        type: number
      salt_g:
        type: number
      sugar_g:
        type: number
      vitamin_b12_mcg:
        type: number
      vitamin_c_mg:
        type: number
      vitamin_d_mcg:
        type: number
      zinc_mg:
        type: number
    type: object
...
  meals.fromPhotoInference:
    properties:
      claude_input_tokens:
        type: integer
      claude_output_tokens:
        type: integer
      confidence:
        type: number
      model:
        type: string
      notes:
        type: string
      original_image_bytes:
        type: integer
      resized_to:
        items:
          type: integer
        type: array
    type: object
  meals.fromPhotoResponse:
    properties:
      inference:
        $ref: '#/definitions/meals.fromPhotoInference'
      meal:
        $ref: '#/definitions/meals.MealEntry'
    type: object
...
  personalrecords.PersonalRecord:
    properties:
      achieved_at:
        type: string
      activity_id:
        type: string
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: string
      pr_type:
        type: string
      unit:
        type: string
      updated_at:
        type: string
      value:
        type: number
    type: object
  personalrecords.upsertRequest:
    properties:
      achieved_at:
        type: string
      activity_id:
        type: string
      external_id:
        type: string
      pr_type:
        type: string
      unit:
        type: string
      value:
        type: number
    type: object
...
info:
  contact: {}
  description: Kazper — personal endurance-fueling and training-coach REST API. All
    endpoints under /products, /meals, and /summary require a bearer token.
  title: Kazper
  version: 0.1.0
paths:
...
  /meals/from_photo:
    post:
      consumes:
      - multipart/form-data
      description: Accepts a `multipart/form-data` upload with an `image` part (JPEG
        or PNG; HEIC is rejected with 415 in v1 because we don't ship libheif). The
        backend resizes the image to a max edge of 1568px, sends it to Claude Vision
        with a tool-forced output, and uses the returned `{name, nutriments_per_100g}`
        to create a freeform meal entry. The response carries both the canonical `meal`
        block and an `inference` block reporting model, confidence, token usage, and
        image dimensions. Returns `503 vision_unavailable` when `ANTHROPIC_API_KEY`
        is not configured.
      parameters:
      - description: Optional client-supplied idempotency key. Hash includes the raw
          image bytes so byte-identical replays return the original meal without a
          second Claude call.
        in: header
        name: Idempotency-Key
        type: string
      - description: Image file (JPEG or PNG; HEIC rejected in v1).
        in: formData
        name: image
        required: true
        type: file
      - description: Meal quantity in grams. Default 100.
        in: formData
        name: quantity_g
        type: number
      - description: RFC 3339 timestamp; default now().
        in: formData
        name: logged_at
        type: string
      - description: breakfast | lunch | dinner | snack.
        in: formData
        name: meal_type
        type: string
      - description: Optional free-text note.
        in: formData
        name: note
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/meals.fromPhotoResponse'
        "400":
          description: image_required | quantity_g_invalid | logged_at_invalid | logged_at_too_far_future
            | meal_type_invalid | note_too_long
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: image_too_large with max_bytes hint
          schema:
            additionalProperties: true
            type: object
        "415":
          description: unsupported_media_type
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: vision_rate_limited with retry_after_seconds
          schema:
            additionalProperties: true
            type: object
        "502":
          description: vision_unexpected_response | vision_response_unparseable
          schema:
            additionalProperties: true
            type: object
        "503":
          description: vision_unavailable (ANTHROPIC_API_KEY not configured)
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: vision_timeout | vision_upstream_error with retry_after_seconds
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Log a meal from a photo via Claude Vision
      tags:
      - meals
  /personal-records:
    get:
      parameters:
      - description: Filter to a single PR type (e.g. 5k)
        in: query
        name: pr_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ personal_records: [...] }'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List personal records
      tags:
      - personal-records
    post:
      consumes:
      - application/json
      description: Creates or updates a personal record, upserting by `external_id`
        (the Garmin PR id). A beaten PR overwrites the prior value in place. Standard
        `Idempotency-Key` header supported. PR values are coaching context only; they
        never feed nutrition computation.
      parameters:
      - description: Optional client-supplied idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Personal record
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/personalrecords.upsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: UPDATE (external_id already present)
          schema:
            $ref: '#/definitions/personalrecords.PersonalRecord'
        "201":
          description: INSERT
          schema:
            $ref: '#/definitions/personalrecords.PersonalRecord'
        "400":
          description: external_id_required | pr_type_required | value_invalid | unit_required
            | achieved_at_required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upsert a personal record (by Garmin PR id)
      tags:
      - personal-records
...
securityDefinitions:
  BearerAuth:
    description: Static bearer token. Use the format `Bearer <token>` with either
      MOBILE_API_TOKEN or AGENT_API_TOKEN.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"